Combine tests/functional/test_install_vcs_git.py with other test modules.
