+++ tests/functional/test_install_vcs.py
@@ -3,6 +3,10 @@ import pytest
 from tests.lib import (
     _change_test_package_version, _create_test_package, pyversion,
 )
+from tests.lib.git_submodule_helpers import (
+    _change_test_package_submodule, _create_test_package_with_submodule,
+    _pull_in_submodule_changes_to_module,
+)
 from tests.lib.local_repos import local_checkout
 
 
@@ -354,3 +358,36 @@ def test_reinstalling_works_with_editible_non_master_branch(script):
     )
     version = script.run('version_pkg')
     assert 'some different version' in version.stdout
+
+
+# TODO(pnasrat) fix all helpers to do right things with paths on windows.
+@pytest.mark.skipif("sys.platform == 'win32'")
+@pytest.mark.network
+def test_check_submodule_addition(script):
+    """
+    Submodules are pulled in on install and updated on upgrade.
+    """
+    module_path, submodule_path = _create_test_package_with_submodule(script)
+
+    install_result = script.pip(
+        'install', '-e', 'git+' + module_path + '#egg=version_pkg'
+    )
+    assert (
+        script.venv / 'src/version-pkg/testpkg/static/testfile'
+        in install_result.files_created
+    )
+
+    _change_test_package_submodule(script, submodule_path)
+    _pull_in_submodule_changes_to_module(script, module_path)
+
+    # expect error because git may write to stderr
+    update_result = script.pip(
+        'install', '-e', 'git+' + module_path + '#egg=version_pkg',
+        '--upgrade',
+        expect_error=True,
+    )
+
+    assert (
+        script.venv / 'src/version-pkg/testpkg/static/testfile2'
+        in update_result.files_created
+    )
